[AArch64] improved support for linux 64 and armv8
Former-commit-id: 32dd8b867934890932c43b9a800be4920d585130

+++ unisim_lib/unisim/util/os/linux_os/aarch64.hh
@@ -275,7 +275,7 @@ namespace linux_os {
     
     static void SetAARCH64SystemCallStatus(LINUX& _lin, int64_t ret, bool error) { SetRegister(_lin, kAARCH64_x0, (parameter_type) ret); }
     
-    void SetSystemCallStatus(int ret, bool error) const { SetAARCH64SystemCallStatus( lin, ret, error ); }
+    void SetSystemCallStatus(int64_t ret, bool error) const { SetAARCH64SystemCallStatus( lin, ret, error ); }
     
     static parameter_type GetSystemCallParam( LINUX& _lin, int id )
     {
